agents/mono/agent-design-patterns-1/requirements.txt
openai>=1.30.0
a2a-sdk[http-server]==0.3.24
uvicorn[standard]>=0.27.0
httpx>=0.27.0
python-dotenv>=1.0.0